chore(lint): enable 17 more golangci linters and fix their 53 findings
Adds canonicalheader, copyloopvar, durationcheck, errorlint, exhaustive,
exptostd, fatcontext, makezero, nilerr, nilnesserr, noctx, nolintlint,
thelper, tparallel, unconvert, usetesting and wastedassign, and lifts the
per-linter issue cap so a batch is never half reported.

No behaviour change. The resumable upload now matches its errors with
errors.As (identical today, nothing wraps them). Deliberate %v on a
secondary error keeps a //nolint with the reason: wrapping would let
exit.For pick that error's exit code (edits begin, install-skills git).
Switches on Op/Status name their ignored cases so a new value forces a
decision; the release status switches fold the ADR-0002 default in.
Test identity checks (verbatim pass-through, memoisation) stay == on
purpose.

Refs #590

+++ commands/subscriptions/prices/convert/convert.go
@@ -90,7 +90,7 @@ func ParseMoney(amount, currency string) (subscriptions.Money, error) {
 	if frac != "" {
 		padded := frac + strings.Repeat("0", 9-len(frac))
 		for _, d := range padded {
-			nanos = nanos*10 + int32(d-'0')
+			nanos = nanos*10 + (d - '0')
 		}
 	}
 	return subscriptions.Money{CurrencyCode: currency, Units: units, Nanos: nanos}, nil
